pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dspy-0to1-guide"
version = "1.0.0"
description = "Comprehensive DSPy tutorial with practical examples and infrastructure integration"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "DSPy 0to1 Guide Contributors", email = "contributors@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"
dependencies = [
    "dspy-ai>=2.4.0",
    "openai>=1.0.0",
    "transformers>=4.30.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",
    "datasets>=2.10.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "prometheus-client>=0.17.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "asyncio-mqtt>=0.13.0",
    "python-dotenv>=1.0.0",
    "jsonlines>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.20.0",
]
ollama = [
    "ollama>=0.1.0",
]
docker = [
    "docker>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/dspy-0to1-guide"
Repository = "https://github.com/yourusername/dspy-0to1-guide"
Documentation = "https://github.com/yourusername/dspy-0to1-guide/tree/main/docs"

[project.scripts]
dspy-setup = "scripts.setup:main"
dspy-evaluate = "scripts.evaluate:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"